requirements.txt
# Database
psycopg2-binary==2.9.9
asyncpg==0.29.0
pgvector==0.2.4

# Embeddings
sentence-transformers==2.3.1
torch==2.1.2
transformers==4.36.2

# LLM
ollama==0.1.6
langchain==0.1.0
langchain-community==0.0.10

# Data Processing
numpy==1.24.3
pandas==2.0.3
datasets==2.16.1

# Monitoring
prometheus-client==0.19.0
psutil==5.9.6

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1
colorama==0.4.6

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1